Move `ty::MainDefinition` to its own little module (cute)
